+++ partiql-planner/src/test/kotlin/org/partiql/planner/HeaderTest.kt
@@ -2,6 +2,7 @@ package org.partiql.planner
 
 import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.Test
+import org.partiql.planner.internal.PartiQLHeader
 
 class HeaderTest {